requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-ai-slim[openai]
python-dotenv
